add command to verify that a cluster is updated
