+++ cmd/action/verify/cluster/created/command.go
@@ -11,8 +11,8 @@ const (
 	short = "Verify if a Tenant Cluster got successfully created."
 	long  = `Check if the Tenant Cluster successfully created.
 
-	* List all Tenant Cluster nodes. Doing so without errors means the apiserver is up.
-	* Check for the "Created" status condition in the AWSCluster CR.
+    * List all Tenant Cluster nodes. Doing so without errors means the apiserver is up.
+    * Check for the "Created" status condition in the AWSCluster CR.
 
 A cluster creation takes up to 30 minutes. This aligns with our cluster creation metric in cluster-operator,
 see https://github.com/giantswarm/cluster-operator/blob/master/service/collector/cluster_transition.go#L135.
