+++ cmd/action/verify/cluster/updated/error.go
@@ -1,6 +1,24 @@
-package ac004
+package updated
 
 import "github.com/giantswarm/microerror"
+
+var invalidConfigError = &microerror.Error{
+	Kind: "invalidConfigError",
+}
+
+// IsInvalidConfig asserts invalidConfigError.
+func IsInvalidConfig(err error) bool {
+	return microerror.Cause(err) == invalidConfigError
+}
+
+var invalidFlagsError = &microerror.Error{
+	Kind: "invalidFlagsError",
+}
+
+// IsInvalidFlags asserts invalidFlagsError.
+func IsInvalidFlags(err error) bool {
+	return microerror.Cause(err) == invalidFlagsError
+}
 
 var wrongClusterStatusConditionError = &microerror.Error{
 	Kind: "wrongClusterStatusConditionError",
