--- cmd/cl002/ac004/zz_generated.flag.go
@@ -1,13 +1,0 @@
-package ac004
-
-import "github.com/spf13/cobra"
-
-type flag struct {
-}
-
-func (f *flag) Init(cmd *cobra.Command) {
-}
-
-func (f *flag) Validate() error {
-	return nil
-}
